Config' file for new genie_example.job
Config' files change the default behaviour of the example job by
assigning new values to variables in the form of var=val pairs.
The use of config' files replaces make_jobs.pl.
Write a new config file for your preferred flavour of job.

Blank lines and comment lines (starting with #) are allowed in
a config file.