Land-to-atmosphere carbon flux is now calculated from the change in land carbon reservoirs over a TRIFFID step, which conserves carbon. This is also the way it's done in IMOGEN.
The UM way of conserving carbon involves carrying over any accumulated
flux into the next TRIFFID period if a reservoir becomes depleted
during the current TRIFFID period.  I don't like this method as it
assumes that the fluxes were physically reasonable through the whole
of the TRIFFID period, which they clearly might not be if the
reservoir was depleted!  Also, my way makes the carbon conservation
check much simpler.