New, cleaner makefile. Goal is to avoid any re-linking if not required.
module libraries (e.g. libfixedseaice.a) and genie.exe
(as a knock-on effect) were re-archived and re-linked
respectively as the .a file was copied into position in
the 'lib' dir.  These mods create the lib in the right
place and so avoid all the extra shinanigans.  This is
better as genie.exe can now be made with debug settings
and not re-linked when running a config file spec'd job.